more private fields
