+++ packages/fuel_rpc/src/metrics.rs
@@ -1,7 +1,5 @@
-use metrics::{Collector, HealthChecker, RegistersMetrics};
+use metrics::{Collector, RegistersMetrics};
 use prometheus::{IntCounter, Opts};
-
-use crate::client::Client;
 
 pub struct Metrics {
     pub fuel_network_errors: IntCounter,
@@ -23,26 +21,5 @@ impl Default for Metrics {
         Self {
             fuel_network_errors,
         }
-    }
-}
-
-impl RegistersMetrics for Client {
-    fn metrics(&self) -> Vec<Box<dyn Collector>> {
-        self.metrics.metrics()
-    }
-}
-
-impl Client {
-    pub fn connection_health_checker(&self) -> HealthChecker {
-        self.health_tracker.tracker()
-    }
-
-    pub(crate) fn handle_network_error(&self) {
-        self.health_tracker.note_failure();
-        self.metrics.fuel_network_errors.inc();
-    }
-
-    pub(crate) fn handle_network_success(&self) {
-        self.health_tracker.note_success();
     }
 }
